Update README.md
Avengers - A movie about superheroes or more like uncontrollable individuals that possess such power, and a god as well are included in the team.

Thor - One of the original of the **avengers** He as well got his own movie and was a god. Son of Odin, and half-brother of **Loki**.

Doctor Strange - A well-known doctor/surgeon that has such talent not only in medical field but in magic as well.

Jurassic World - There's a tour in an island theme park populated by dinosaurs created from prehistoric DNA. Era wherein dinosaur still exist by getting their fossils and creating them again. 

Transformer - A huge humanoid-robots that came from another planet called "**Cybertron**". A technological world/planet that is full of shining metals and robots which called transformers sooner called "**Autobots" and "Decepticon**".