Create README.md
Avengers - A group of individuals that has different kind of powers, they grouped themselves together even though they don't want to do it but in order to save the planet Earth, they should do it. The group consists of billionaire, an archer, spy, enraged green beast, a lightning god and an old human that wears America outfit.

Thor - One of the original member of the group called "Avenger" of planet Earth. A god that can control a lightning, and son of Odin ( The father of all ), a brother of Loki ( God of Mischief ).

Doctor Strange - A well-known doctor/surgeon. Best of the best skilled surgeon in the world. But due to some accident, he lost his ability to save someone life. He lost his ability as a surgeon but a new power/skill came and that's what makes who he is right now.

Jurassic World - An island theme park populated by dinosaurs created from prehistoric DNA. Dinosaurs came back from fossil and many things happened.

Transformer - A huge humanoid-robots that came from another planet called "Cybertron". The robots can transform into different kind of cars that exists in the modern era of the world.